Move UTXOs definition from primary to primary/common
