+++ wallet/subnet/primary/api.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/ava-labs/coreth/ethclient"
 	"github.com/ava-labs/coreth/plugin/evm"
-	"github.com/ethereum/go-ethereum/common"
 
 	"github.com/ava-labs/avalanchego/api/info"
 	"github.com/ava-labs/avalanchego/codec"
@@ -24,6 +23,9 @@ import (
 	"github.com/ava-labs/avalanchego/wallet/chain/c"
 	"github.com/ava-labs/avalanchego/wallet/chain/p"
 	"github.com/ava-labs/avalanchego/wallet/chain/x"
+
+	walletcommon "github.com/ava-labs/avalanchego/wallet/subnet/primary/common"
+	ethcommon "github.com/ethereum/go-ethereum/common"
 )
 
 const (
@@ -60,7 +62,7 @@ type AVAXState struct {
 	XCTX    x.Context
 	CClient evm.Client
 	CCTX    c.Context
-	UTXOs   UTXOs
+	UTXOs   walletcommon.UTXOs
 }
 
 func FetchState(
@@ -91,7 +93,7 @@ func FetchState(
 		return nil, err
 	}
 
-	utxos := NewUTXOs()
+	utxos := walletcommon.NewUTXOs()
 	addrList := addrs.List()
 	chains := []struct {
 		id     ids.ID
@@ -143,13 +145,13 @@ func FetchState(
 
 type EthState struct {
 	Client   ethclient.Client
-	Accounts map[common.Address]*c.Account
+	Accounts map[ethcommon.Address]*c.Account
 }
 
 func FetchEthState(
 	ctx context.Context,
 	uri string,
-	addrs set.Set[common.Address],
+	addrs set.Set[ethcommon.Address],
 ) (*EthState, error) {
 	path := fmt.Sprintf(
 		"%s/ext/%s/C/rpc",
@@ -161,7 +163,7 @@ func FetchEthState(
 		return nil, err
 	}
 
-	accounts := make(map[common.Address]*c.Account, addrs.Len())
+	accounts := make(map[ethcommon.Address]*c.Account, addrs.Len())
 	for addr := range addrs {
 		balance, err := client.BalanceAt(ctx, addr, nil)
 		if err != nil {
@@ -188,7 +190,7 @@ func FetchEthState(
 // expires, then the returned error will be immediately reported.
 func AddAllUTXOs(
 	ctx context.Context,
-	utxos UTXOs,
+	utxos walletcommon.UTXOs,
 	client UTXOClient,
 	codec codec.Manager,
 	sourceChainID ids.ID,
